pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = 'mlebench'
readme = "README.md"
requires-python = ">=3.11"
version = "1.0.0"
dependencies = [
  "appdirs>=1.4.4",
  "docker>=7.1",
  "fastparquet>=2024.5.0",
  "kaggle>=1.6,<1.7",
  "levenshtein>=0.25.1",
  "openai>=1.10.0",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "py7zr>=0.21",
  "scikit-learn>=1.5",
  "pillow>=10.3",
  "python-dotenv>=1.0.1",
  "aiohttp>=3.9.5",
  "tenacity>=8.5.0",
  "nbformat>=5.10.4",
  "nbconvert>=7.16.4",
  "ipython>=8.26.0",
  "pymongo>=4.8.0",
  "pycocotools>=2.0.8",
  "diskcache>=5.6.3",
  "pyquaternion>=0.9.9",
  "shapely>=2.0.5",
  "smolagents>=1.20.0",
  "matplotlib>=3.10.3",
  "seaborn>=0.13.2",
]

[tool.setuptools]
packages = { find = { include = ["mlebench", "mlebench.*"] } }

[project.scripts]
mlebench = "mlebench.cli:main"

[project.optional-dependencies]
formatters = ["black>=23.3", "isort>=5.13"]
dev = ["pytest>=8.2", "mypy>=1.10", "pre-commit>=3.7"]

[tool.setuptools.package-data]
mlebench = ["competitions/**/*"]

[tool.black]
line-length = 100 # match the precommit

[tool.pytest.ini_options]
# Pytest can take a long time to collect test cases
# when many Kaggle competitions have been downloaded.
# We avoid this up by telling Pytest where the
# tests live ahead of time.
testpaths = ["tests"]

